backend/requirements.txt
Django = 3.1.3
django-cors-headers   =  3.5.0
django-templated-mail = 1.1.1
djangorestframework   =  3.12.2
djangorestframework-simplejwt 4.5.0
djoser  =  2.1.0
Pillow  =  8.0.1
pip     =  20.2.4
Jinja2  =  2.11.2
